+++ compiler-tests/src/test/data/dump/ir/dependencygraph/PrimitiveProviderFactoryParamUsesPrimitiveFactory.kt
@@ -1,0 +1,12 @@
+@DependencyGraph
+interface AppGraph {
+  val int: Int
+  val intProvider: () -> Int
+  val longProvider: () -> Long
+  val longProvider2: () -> Long
+
+  @DependencyGraph.Factory
+  interface Factory {
+    fun create(@Provides int: Int, @Provides long: Long): AppGraph
+  }
+}