+++ compiler-tests/src/test/data/dump/ir/provides/ProviderInliningDisabled.kt
@@ -1,0 +1,6 @@
+// ENABLE_PROVIDER_INLINING: false
+
+@BindingContainer
+object Bindings {
+  @Provides fun provideInt(): Int = 3
+}